Move subpart/subjgrp logic to distinct module
Also rewrites touched tests w/ py.test
